requirements.txt
fastapi
uvicorn
piper-tts
python-multipart
pyotp
python-jose[cryptography]